improve --extend
