+++ pr_agent/cli.py
@@ -19,13 +19,21 @@ For example:
 - cli.py --pr_url=... reflect
 
 Supported commands:
-review / review_pr - Add a review that includes a summary of the PR and specific suggestions for improvement.
-ask / ask_question [question] - Ask a question about the PR.
-describe / describe_pr - Modify the PR title and description based on the PR's contents.
-improve / improve_code - Suggest improvements to the code in the PR as pull request comments ready to commit.
-reflect - Ask the PR author questions about the PR.
-update_changelog - Update the changelog based on the PR's contents.
+-review / review_pr - Add a review that includes a summary of the PR and specific suggestions for improvement.
 
+-ask / ask_question [question] - Ask a question about the PR.
+
+-describe / describe_pr - Modify the PR title and description based on the PR's contents.
+
+-improve / improve_code - Suggest improvements to the code in the PR as pull request comments ready to commit.
+Extended mode ('improve --extended') employs several calls, and provides a more thorough feedback
+
+-reflect - Ask the PR author questions about the PR.
+
+-update_changelog - Update the changelog based on the PR's contents.
+
+
+Configuration:
 To edit any configuration parameter from 'configuration.toml', just add -config_path=<value>.
 For example: 'python cli.py --pr_url=... review --pr_reviewer.extra_instructions="focus on the file: ..."'
 """)
